Minor refactoring.

+++ libdf3d/base/Common.cpp
@@ -10,7 +10,7 @@ glm::vec3 jsonGetVec3(const Json::Value &root, glm::vec3 defVal)
         return defVal;
     }
 
-    return glm::vec3((float)root[0u].asDouble(), (float)root[1u].asDouble(), (float)root[2u].asDouble());
+    return glm::vec3(root[0u].asFloat(), root[1u].asFloat(), root[2u].asFloat());
 }
 
 glm::vec4 jsonGetVec4(const Json::Value &root, glm::vec4 defVal)
@@ -20,7 +20,7 @@ glm::vec4 jsonGetVec4(const Json::Value &root, glm::vec4 defVal)
         return defVal;
     }
 
-    return glm::vec4((float)root[0u].asDouble(), (float)root[1u].asDouble(), (float)root[2u].asDouble(), (float)root[3u].asDouble());
+    return glm::vec4(root[0u].asFloat(), root[1u].asFloat(), root[2u].asFloat(), root[3u].asFloat());
 }
 
 DF3D_DLL std::string glmVecToString(const glm::vec3 &v)
